Move PasswordRules to more common module and remove duplicated code
