+++ control-admin/src/main/java/fi/nls/oskari/control/admin/UsersHandler.java
@@ -14,6 +14,7 @@ import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.oskari.log.AuditLog;
+import org.oskari.user.util.PasswordRules;
 
 import java.util.List;
